Rename experiment_save to just experiment since we are also loading models from here
