+++ AIToolbox/cloud/GoogleCloud/model_save.py
@@ -2,7 +2,7 @@ import os
 from google.cloud import storage
 
 from AIToolbox.cloud.AWS.model_save import KerasS3ModelSaver, TensorFlowS3ModelSaver, PyTorchS3ModelSaver
-from AIToolbox.experiment_save.local_save.local_model_save import KerasLocalModelSaver, TensorFlowLocalModelSaver, PyTorchLocalModelSaver
+from AIToolbox.experiment.local_save.local_model_save import KerasLocalModelSaver, TensorFlowLocalModelSaver, PyTorchLocalModelSaver
 
 
 class BaseModelGoogleStorageSaver:
