+++ AIToolbox/cloud/GoogleCloud/results_save.py
@@ -2,7 +2,7 @@ import os
 from google.cloud import storage
 
 from AIToolbox.cloud.AWS.results_save import S3ResultsSaver
-from AIToolbox.experiment_save.local_save.local_results_save import LocalResultsSaver
+from AIToolbox.experiment.local_save.local_results_save import LocalResultsSaver
 
 
 class BaseResultsGoogleStorageSaver:
