+++ AIToolbox/experiment/local_experiment_saver.py
@@ -1,9 +1,9 @@
 import datetime
 import time
 
-from AIToolbox.experiment_save.experiment_saver import AbstractExperimentSaver
-from AIToolbox.experiment_save.local_save.local_model_save import PyTorchLocalModelSaver, KerasLocalModelSaver, AbstractLocalModelSaver
-from AIToolbox.experiment_save.local_save.local_results_save import LocalResultsSaver
+from AIToolbox.experiment.experiment_saver import AbstractExperimentSaver
+from AIToolbox.experiment.local_save.local_model_save import PyTorchLocalModelSaver, KerasLocalModelSaver, AbstractLocalModelSaver
+from AIToolbox.experiment.local_save.local_results_save import LocalResultsSaver
 
 
 class BaseFullExperimentLocalSaver(AbstractExperimentSaver):
