+++ tests/test_cloud/test_AWS/test_model_load.py
@@ -1,7 +1,7 @@
 import unittest
 
 from AIToolbox.cloud.AWS.model_load import PyTorchS3ModelLoader
-from AIToolbox.experiment_save.local_load.local_model_load import AbstractLocalModelLoader, PyTorchLocalModelLoader
+from AIToolbox.experiment.local_load.local_model_load import AbstractLocalModelLoader, PyTorchLocalModelLoader
 
 
 class TestPyTorchS3ModelLoader(unittest.TestCase):
