+++ tests/test_experiment_save/test_local_save/test_local_model_save.py
@@ -4,7 +4,7 @@ import shutil
 
 from tests.utils import *
 
-from AIToolbox.experiment_save.local_save.local_model_save import *
+from AIToolbox.experiment.local_save.local_model_save import *
 from AIToolbox.torchtrain.train_loop import TrainLoop
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
