+++ tests/test_torchtrain/test_callbacks.py
@@ -6,9 +6,9 @@ from AIToolbox.torchtrain.callbacks.callbacks import AbstractCallback, EarlyStop
 from AIToolbox.torchtrain.callbacks.model_save_callbacks import ModelCheckpoint, ModelTrainEndSave
 from AIToolbox.torchtrain.train_loop import TrainLoop
 from AIToolbox.cloud.AWS.model_save import PyTorchS3ModelSaver
-from AIToolbox.experiment_save.local_save.local_model_save import PyTorchLocalModelSaver
-from AIToolbox.experiment_save.experiment_saver import FullPyTorchExperimentS3Saver
-from AIToolbox.experiment_save.local_experiment_saver import FullPyTorchExperimentLocalSaver
+from AIToolbox.experiment.local_save.local_model_save import PyTorchLocalModelSaver
+from AIToolbox.experiment.experiment_saver import FullPyTorchExperimentS3Saver
+from AIToolbox.experiment.local_experiment_saver import FullPyTorchExperimentLocalSaver
 
 
 class TestAbstractCallback(unittest.TestCase):
